# services/api/src/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Hill90 API
  description: Hill90 platform API — agent management and lifecycle operations.
  version: 0.1.0
servers:
  - url: /
    description: Relative (behind Traefik reverse proxy)
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /provider-connections:
    get:
      summary: List provider connections
      description: Returns provider connections owned by the authenticated user. API keys are never exposed.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Connection list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    name:
                      type: string
                    provider:
                      type: string
                    api_base_url:
                      type: string
                      nullable: true
                    is_valid:
                      type: boolean
                      nullable: true
                    created_at:
                      type: string
                      format: date-time
                    updated_at:
                      type: string
                      format: date-time
        "401":
          description: Not authenticated
        "403":
          description: Requires user role

    post:
      summary: Create provider connection
      description: >
        Creates a new provider connection with an encrypted API key.
        The key is encrypted at rest and never returned in responses.
        Admins can set platform=true to create a platform connection (created_by=NULL).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - provider
                - api_key
              properties:
                name:
                  type: string
                provider:
                  type: string
                  description: Provider identifier (e.g., openai, anthropic)
                api_key:
                  type: string
                  description: Provider API key (encrypted at rest, never returned)
                api_base_url:
                  type: string
                  description: Custom API base URL (optional)
                platform:
                  type: boolean
                  description: Admin-only. If true, creates a platform connection (created_by=NULL) accessible to all platform models.
      responses:
        "201":
          description: Connection created
        "400":
          description: Invalid input
        "401":
          description: Not authenticated
        "403":
          description: Requires user role
        "409":
          description: Duplicate connection name
  /provider-connections/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    put:
      summary: Update provider connection
      description: Updates a provider connection. If api_key is provided, it is re-encrypted and is_valid is reset.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                provider:
                  type: string
                api_key:
                  type: string
                api_base_url:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Connection updated
        "401":
          description: Not authenticated
        "403":
          description: Requires user role
        "404":
          description: Connection not found
        "409":
          description: Duplicate connection name

    delete:
      summary: Delete provider connection
      description: Deletes a provider connection and atomically cascades to router models referencing it (JSONB route cleanup). Single models are cascaded via FK ON DELETE CASCADE.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Connection deleted
        "401":
          description: Not authenticated
        "403":
          description: Requires user role
        "404":
          description: Connection not found
        "500":
          description: Cascade cleanup failed — connection delete rolled back
  /provider-connections/{id}/validate:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Validate provider connection
      description: >
        Tests whether the provider key is valid by sending a lightweight request
        through the AI service. Updates is_valid on the connection.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  is_valid:
                    type: boolean
                  error:
                    type: string
                    description: Error message if validation failed
        "401":
          description: Not authenticated
        "403":
          description: Requires user role
        "404":
          description: Connection not found
  /provider-connections/{id}/models:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: List provider models
      description: >
        Returns available models from the provider API for a given connection.
        Each model includes auto-detected type and capabilities.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Provider model list (may include error field on failure)
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: Provider-prefixed model ID (e.g., openai/gpt-4o)
                        display_name:
                          type: string
                        detected_type:
                          type: string
                          enum: [chat, embedding, audio, image, transcription]
                        capabilities:
                          type: array
                          items:
                            type: string
                  provider:
                    type: string
                  error:
                    type: string
                    description: Error message if model listing failed
        "401":
          description: Not authenticated
        "403":
          description: Requires user role
        "404":
          description: Connection not found
...